requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3

# Database drivers
psycopg2-binary
PyMySQL==1.1.0

# Data processing and analytics
pandas==2.1.3
numpy
sqlalchemy==2.0.23
openpyxl==3.1.2
xlrd==2.0.1
python-dateutil==2.8.2
scipy==1.11.4
scikit-learn==1.3.2

# Background tasks and caching
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
flower==2.0.1

# Authentication and security
django-allauth==0.57.0
PyJWT==2.8.0
cryptography>=41.0.7
bcrypt==4.1.2
django-security
django-ratelimit==4.1.0
django-axes==6.1.1
social-auth-app-django==5.4.0
django-oauth-toolkit==1.7.1

# API and serialization
marshmallow==3.20.1
jsonschema==4.19.2
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.12.1

# File processing and export
reportlab==4.0.6
Pillow==10.1.0
python-magic==0.4.27
xlsxwriter==3.1.9
unicodecsv==0.14.1
python-docx==1.1.0

# Web scraping and HTTP
requests==2.31.0
urllib3==2.2.2
httpx==0.25.2

# Email and communications
django-sendgrid-v5==1.2.3
twilio==8.10.0

# Logging and monitoring
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0
django-silk==5.0.4
django-request-logging==0.7.5
raven==6.10.0
psutil==5.9.6
structlog==23.2.0
statsd==4.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
responses==0.24.1
freezegun==1.2.2

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0
invoke==2.2.0
ipython==8.17.2
django-livereload-server==0.4

# Environment management
python-dotenv==1.0.0
python-decouple==3.8
django-environ==0.11.2
dynaconf==3.2.4

# Time and timezone handling
pytz==2023.3
arrow==1.3.0
pendulum==3.0.0
maya==0.6.1

# Data validation and serialization
cerberus==1.3.5
pydantic==2.7.0
dill==0.3.7
voluptuous==0.14.1

# Data visualization support
matplotlib==3.8.2
seaborn==0.13.0

# Workflow and state management
django-fsm==2.8.1

# Search functionality
django-haystack==3.2.1
whoosh==2.7.4

# Caching and performance
django-cachalot==2.6.1
memory-profiler==0.61.0
django-bulk-update==2.2.0
django-model-utils==4.3.1

# Additional utils
python-slugify==8.0.1
humanize==4.8.0
babel==2.13.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
pathlib2==2.3.7.post1
click==8.1.7
rich==13.7.0
pyyaml==6.0.1
toml==0.10.2
orjson==3.9.10
furl==2.1.3
tenacity==8.2.3
tqdm==4.66.1

# Channels (WebSocket)
channels==4.0.0
channels-redis==4.1.0

# Additional admin interface enhancements
django-admin-interface==0.28.6
django-colorfield==0.10.1

# Security enhancements (install django-security separately if needed)
# django-security can be installed later if required and compatible.

gunicorn==21.2.0

# Commented out problematic or unused packages:
# gunicorn==21.2.0       # Not for Windows dev
# uwsgi==2.0.22          # Not for Windows dev
# django-db-pool==1.0.3  # Not found on PyPI currently
# django-contrib-gis      # Uncomment if GIS features needed
# webdavclient3==3.14.6  # Uncomment if WebDAV needed
# pillow-heif==0.13.0    # Uncomment if HEIF support needed

